# server/api/openapi.yaml
openapi: 3.0.0
info:
  title: Consciousness API
  version: 1.0.0
servers:
  - url: /
paths:
  /health:
    get:
      summary: Health check endpoint
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  backend:
                    type: string
  /realities:
    post:
      summary: Create an encoded reality
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                  maxLength: 80
                data:
                  type: object
              required:
                - id
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  details:
                    type: array
                    items:
                      type: object
    get:
      summary: List all encoded realities (Not implemented)
      responses:
        '501':
          description: Not implemented
  /realities/{id}:
    get:
      summary: Get an encoded reality by id
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
        '404':
          description: Not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string